changed style to seperate class
